unpopulated-alloc.c: Drop restriction for DMA_BIT_MASK(64)
The single at the monent in-tree user of that API sets DMA_BIT_MASK(64),
so this is why this restriction is here.

But there is an out-of-tree user (PVRKM) which sets DMA_BIT_MASK(40),
so drop the restriction for now.
In general it is not good to make an excuse for out-of-tree modules...

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>