net: renesas: rswitch: do not set IRQ delay for TSNs
IRQ delays were introduced because without them VMQs perform
a lot of interrupts. This mechanism is needed for VMQs to reduce
the amount of interrupts and as a result, improve performance.
On the other hand, any delays, while using TSNs lead to bandwidth
reduction, so in such cases delay should be zeroed.

Signed-off-by: Leonid Komarianskyi <leonid_komarianskyi@epam.com>
Reviewed-by: Dmytro Firsov <dmytro_firsov@epam.com>